bugfix: fix data_ready() check, avoid bogus 'server is building..'
graphs_all, targets_all and even metrics can be empty, depending
on configuration.  that doesn't mean the data is not ready.

this fixes a bug where the app would say "server is building
datastructures" just because graphs_all or targets_all was empty,
which was especially nasty in combination with the bug that's fixed
in the previous commit.