remove configure function(s) from finished target
once the target is created and configured, there's no need to
keep those around.  this allows pickling targets (and targets_all)
or using a real metrics/tags database (after some work..)
this may also have a small positive impact on resource usage.
the only place this was ever used after generating the targets
was in the inspect view, so now you can longer see the exact
target config used to build the target, that's a small sacrifice.