remove plugin loading&information from app
we want to isolate everything related to generating the enhanced
metrics from the web app, which should be just a consumer, and
unaware of how the metrics are generated.  this will allow
different backends, and ultimately, native tags in graphite itself
* from inspect view, remove used plugins, not very useful anyway
* from debug page, remove plugin shortcuts section. now you just
need to know how to use the query interface.