make tooltips work
bootstrap needs to be loaded after jquery for tooltips to work.
note: this means we load jquery and bootstrap twice in GE.
the alternative is that we remove the whole
templates/snippet.graph-deps.tpl and just add all unique
dependencies to templates/page.tpl;
but then embeddability is endangered/broken:
graph html should load all its JS so that you can
easily embed it anywhere.