automatic unit derivation
When an explicit "unit=" term with a per-unit-time denominator is
included in a GEQL query, graph-explorer will now widen the search to
include data sources with that unit /or/ any other unit that appears to
be compatible with it /or/ any unit compatible with the numerator unit.
If a datasource is found matching the numerator unit, Graph-explorer
will automatically apply a derivative function to change it into a rate.

Data sources with target_type='counter' will be automatically graphed as
a derivative, to show a rate, unless a contradicting explicit "unit="
term was included in the query.

The conversion machinery here is also capable of applying an integral
transformation to get from "unit/time" to "unit", but this is currently
not enabled as (a) I'm not sure if it would be confusing, and (b) I'm
not sure it would be useful, and (c) it would probably make most ES
queries slower with all the extra units it would search for.

Tests extended to allow for the new behavior.