refactor query compilation and matching
Most importantly, this changes the way that the query AST is represented
between GEQL compilation and ES query construction. The
Query.compile_patterns() method (now Query.compile_pattern(), since it
yields a single pattern tree) creates a tree of tuples where the first
element of the tuple is the name of the operation to be performed, and
the remaining elements are arguments to that operation. Arguments may be
sub-tuples representing more operations. For an example, see the
method's docstring.

This also adds to the possible operations that can be represented. Since
the GEQL behavior for "tag=val", "tag=", and "=val" are all fairly
different, I've given them different operation names at compile time so
that later evaluation or ES query construction can be dispatched more
easily based only on the op name. Same with "tag:val", "tag:", and
":val". "negate" is split out into its own operation as well. "And" and
"or" operations are added, as they'll soon be useful.

A Query's 'compiled_patterns' attribute (now 'compiled_pattern') no
longer associates compiled query terms with the original GEQL term from
the user, since it was somewhat messy to do so and I didn't really like
the "if you have twice the exact same "word" (ignoring leading '!'), the
last one wins" behavior.

Tests were adjusted to match the new query AST structure.