assure that targets always have 'tags' set
historically at this point we didn't care so much about
the tags attribute anymore, as going forward we only needed
variables and constants to do the graph rendering.

but now this is important for target_modifiers to work that need to
inspect tags, such as the unit tag.
(we could also have them search in constants and variables but
that seems clunky, we could also return tags dynamically in the
Target object based on variables/constants but
this solution seems more clear)