implement equivalence for sum by. fixes #89
diamond provides metrics like cpu.<core number>.<type>
where <core number> can be 'total' (sum of all cores)
collectd only provides the individuals

we need to be able to see total cpu usage from both plugins,
whether or not you (only) have the per-core metrics and/or the totals,
preferring only the total if available.

this is achieved by having the total metric setting core=_sum_
and summing by core. for every tag being summed by,
the algorithm checks if there's one that has value <tag>=_sum_
and if so uses that (only) and removes the other ones.

there's some more intricate details about how it works, but that'll be
for the docs